Implement configurable flag prefix characters and flag separator.
It is now possible to set up a parser to recognize certain characters
as flag prefixes (instead of the default dash). With that goes the flag
separator, which tells the parser to treat all following command line
words as positional arguments. (That is usually a double dash).

There are some challenges with the current implementation. For example,
it is not checked if newly created flags match exactly the flag
separator.

Lots more unit testing needs to be done still.