Change implementation of flag prefix characters and flag separators.
Flag prefix characters are now always stored in an ArgumentParser,
even if they are left default. The flag separator is also stored even
if deafult, unless it was explicitly disabled.

This simplifies logic in some function, because they no longer need
to create an ad-hoc string for either of these values if they are
missing. It also simplifies prohibiting the creation of flags that would
be identical to the flag separator.