Change parsign error reporting
ParsingResult returned from cap_parser_parse_noexit now contains
an error code to identify the error type to the caller. Error messages
are created entirely by the caller (i.e. cap_parser_parse). The mSuccess
member is no longer necessary. Successful parsing is now indicated
by the PER_NO_ERROR error type.

In some cases, extra words need to be printed as part of the error
message, such as the name of a flag. These strings are passed in the
mFirstErrorWord and mSecondErrorWord members of the ParsingResult
structure.