Add more descriptive and customizable display informations
Added customization options for name of the program, values of arguments
and values of flags, if applicable. These values are used for printing
a more presentable usge message (and a help message in the future).

Improved the slicer tool to handle #if and #endif directives more
carefully. Only the outer-most one #if-#endif block is ignored (to get
rid of include guards). This change is required to enable checking
the compilation platform using #if _WIN32 when guessing the program name
based on the first command line argument.