Implement an automatic help flag.
The help flag heeds to be recognized directly by the parser. If it were
up to the user to create one manually the parser would cause an error
when only the help flag is present and none of the required arguments,
for example.

Also added a new parser factory: cap_parser_make_default. It behaves
the same as make_empty but additionally configures a default
help flag '-h'.