WIP: refactoring parser
Partial refactoring of the main parsing function - cap_parser_parse_noexit.
The goal is to split it into multiple logical parts to better follow Single
Responsibility. It would also allow clearer control flow (using function
calls instead of goto.)

Current problem is separating a function that handles flags. The parser
handles the flag separator and the help flag largely as normal flags.
It is, however, necessary to distinguish the two special flags from the
others and implementation as of now does not do that well - the function
_cap_parser_parse_one_flag checks for those special cases (to avoid
registering them as regular flags), then the main function checks again
to change its state.

See comment on line 1173.