WIP slicer v2: partial parsing of functions
Current progress:
- extend parser to recognize type definitions (enums, structs, typedefs)
  - typedef only works for structures and enums

TODO:
- parser should recognize #include and #if directives, this will make
  it easier to reproduce files
- implement function and type reproducers
  - given a file, clone a type or function declaration or definition
    into another file, optionally making it static
  - should reuse a lot of code from parser

The objective is to change the distribution format of the library to
one header file + one source file (instead of the current single header.)
The header file should declare only those symbols that the user should
have access to. All other symbols should be static in the source file.
Users of the library would then compile this source file and link it with
their program.

This involves several tasks:
- reimplement current header files to follow the usual header+source format
- reimplement the slicer tool to take these header+source files and
  turn them into a header (containing public symbols) and a large source
  file containing all function implementatins.
- adapt the test system to use the new format
- change README.md accordingly