WIP slicer version 2
The objective is to change the distribution format of the library to
one header file + one source file (instead of the current single header.)
The header file should declare only those symbols that the user should
have access to. All other symbols should be static in the source file.
Users of the library would then compile this source file and link it with
their program.

This involves several tasks:
- reimplement current header files to follow the usual header+source format
- reimplement the slicer tool to take these header+source files and
  turn them into a header (containing public symbols) and a large source
  file containing all function implementatins.
- adapt the test system to use the new format
- change README.md accordingly

Current progress:
- implemented topological ordering of header files
  - function topsort_headers in the header_analysis module
  - this reorders registered headers in the array populated by
    register_headers so that no header file references another that is
    present further to the right in the array
  - it also invalidates information about file inclusions stored in
    the include member of struct Header (this is fine because that
    information exists solely to allow this reordering and is not useful
    afterwards)