WIP slicer version 2
The objective is to change the distribution format of the library to
one header file + one source file (instead of the current single header.)
The header file should declare only those symbols that the user should
have access to. All other symbols should be static in the source file.
Users of the library would then compile this source file and link it with
their program.

This involves several tasks:
- reimplement current header files to follow the usual header+source format
- reimplement the slicer tool to take these header+source files and
  turn them into a header (containing public symbols) and a large source
  file containing all function implementatins.
- adapt the test system to use the new format
- change README.md accordingly

Current progress:
- moved slicer code into slicer_v2 dir
- created basic tools and structures for header analysis
  - system header inclusions (for all headers together)
  - user header inclusions (for each header separately)