Implement flag aliases
Flag aliases allow setting up an alternate name for a flag, e.g. having
a long and short version of it. Flag aliases are handled the same way
as the main flag names on the command line. Most importantly, any values
created for a flag alias are stored under the main flag name, not the
alias.