Re-implement storage of flag and positional values
Since arguments can now be variadic, they need to store values in a very
similar way to flags, so it makes sense to use the same data structure
for both. Flag and positional values are now stored as an array
containing NamedValues objects. NamedValues contains a string name and
an array of TypedUnion values.

This structure could be further optimized for example by having
a decision-tree-like structure which would map names to their
NamedValues objects. There surely exist other improvements.