wip doxygen: add a lot of documentation to parser
Started a topic describing the ArgumentParser type and its functions.
About half of the documentation is done there. It really seems useful
to create sub-topics in the main topics (parser especially), since those
pages are becoming very long now.

Moved the Quick Start section from README.md to a new file, so that the
quick start guide can be used in the doxygen docs as well. Started a
documentation page about building the library, but it is very short
right now.

Changed some doxygen configuration, such as generating a treeview. I
should possibly look into modifying the layout of the generated pages
and other customizations.

The goal is to use GitHub actions to generate this documentation
and deploy it to GitHub Pages with each push to main.