wip: generate documentation with doxygen
Started using doxygen for generating documentation. Need to add data
structures and related functions to reasonable groups (topics) and
perhaps create sub-topics within those topics. Also should look into
modifying the layout of the pages and other customizations.

The goal is to use GitHub actions to generate this documentation
and deploy it to GitHub Pages with each push to main.